Plugin compiles and installs with Boost dependency in CMakeLists. Rename files to match repository name.
