+++ StainAnalysis-plugin.h
